+++ app/src/main/java/jp/panta/misskeyandroidclient/ui/explore/ExploreFragment.kt
@@ -30,10 +30,10 @@ import kotlinx.coroutines.FlowPreview
 import net.pantasystem.milktea.common.ResultState
 import net.pantasystem.milktea.common.StateContent
 import net.pantasystem.milktea.model.user.query.*
-import net.pantasystem.milktea.user.ToggleFollowViewModel
 import net.pantasystem.milktea.user.UserCardActionHandler
-import net.pantasystem.milktea.user.UserDetailCard
-import net.pantasystem.milktea.user.UserDetailCardAction
+import net.pantasystem.milktea.user.compose.UserDetailCard
+import net.pantasystem.milktea.user.compose.UserDetailCardAction
+import net.pantasystem.milktea.user.viewmodel.ToggleFollowViewModel
 
 @FlowPreview
 @ExperimentalCoroutinesApi
