--- app/src/main/res/layout/item_clickable_simple_user.xml
@@ -1,64 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<layout xmlns:tools="http://schemas.android.com/tools"
-        xmlns:app="http://schemas.android.com/apk/res-auto"
-        xmlns:android="http://schemas.android.com/apk/res/android">
-    <data>
-
-        <variable
-                name="user"
-                type="net.pantasystem.milktea.common_viewmodel.UserViewData" />
-
-
-        <import type="net.pantasystem.milktea.common_android.ui.SafeUnbox" />
-    </data>
-    <androidx.cardview.widget.CardView android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-
-            android:layout_marginTop="0.5dp"
-            app:cardBackgroundColor="?attr/colorSurface"
-            app:cardCornerRadius="0dp"
-            >
-        <RelativeLayout
-                android:layout_width="match_parent" android:layout_height="match_parent"
-                android:padding="8dp"
-                app:user="@{user.user}"
-                app:mainNameView="@{nameMain}"
-                app:subNameView="@{nameSub}"
-                app:transitionDestinationUser="@{user.user}"
-                >
-            <ImageView
-                    android:id="@+id/avatarIcon"
-                    android:layout_width="46dp"
-                    android:layout_height="46dp"
-                    android:layout_marginEnd="8dp"
-                    tools:src="@drawable/ic_person_add_black_24dp"
-                    tools:ignore="ContentDescription"
-                    app:circleIcon="@{user.user.avatarUrl}"
-                    app:transitionDestinationUser="@{user.user}"
-                    />
-            <TextView
-                    android:id="@+id/nameMain"
-                    android:layout_toEndOf="@id/avatarIcon"
-                    android:layout_width="match_parent"
-                    android:layout_height="wrap_content"
-                    android:singleLine="true"
-                    tools:text="\@userName"
-                    android:textSize="16sp"
-                    android:layout_alignParentEnd="true"
-                    android:layout_marginBottom="8dp"
-                    />
-            <TextView
-                    android:id="@+id/nameSub"
-                    android:layout_width="match_parent"
-                    android:layout_height="wrap_content"
-                    android:layout_toEndOf="@id/avatarIcon"
-                    android:layout_below="@id/nameMain"
-                    tools:text="name"
-                    android:layout_alignParentEnd="true"
-
-                    />
-
-        </RelativeLayout>
-    </androidx.cardview.widget.CardView>
-
-</layout>
