+++ modules/features/user/src/main/java/net/pantasystem/milktea/user/viewmodel/SelectedUserViewModel.kt
@@ -1,4 +1,4 @@
-package net.pantasystem.milktea.user.selectable
+package net.pantasystem.milktea.user.viewmodel
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
